Organize API into Blueprints, Fix SITL for Prod UAV Handler, Fix bugs
